make range checking the default

+++ matrix/test.c
@@ -18,7 +18,7 @@ int main (void)
 
   for (i = 0 ; i < N ; i++) {
     for (j = 0 ; j < M ; j++) {
-      k++ ; gsl_matrix_set(m,i,j,k) ;
+      k++ ; gsl_matrix_set(m,i,j,(double)k) ;
     }
   }
